+++ src/_/colors.ts
@@ -1,17 +1,17 @@
-import Color from 'color';
+import color_ from 'color';
 import * as colorNames from 'color-name';
 
-import type Color_ from './color.types.js';
+import Color, { ColorValue } from './color.types.js';
 
-export type { default as ColorValue } from 'color';
+export type { ColorValue } from './color.types.js';
 
 export type ColorName = keyof typeof colorNames;
 
-type ColorPlus = typeof Color_ & {
-  fromName(colorName: ColorName): Color;
+type ColorPlus = typeof Color & {
+  fromName(colorName: ColorName): ColorValue;
 };
 
-export const color = Color as ColorPlus;
+export const color = color_ as ColorPlus;
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 color.prototype.getName = color.prototype.toString;
@@ -21,15 +21,15 @@ export const hsl = color.hsl;
 
 // Patch color's lossy hex method.
 // See: https://github.com/Qix-/color/issues/243
-const oldHex = (color.prototype as Color).hex;
-(color.prototype as Color).hex = function (value) {
+const oldHex = (color.prototype as ColorValue).hex;
+(color.prototype as ColorValue).hex = function (value) {
   if (arguments.length > 0) {
-    return new Color(value);
+    return new color(value);
   }
   // @ts-expect-error  (@types/color don't fully match color's actual implementation)
   // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
   return this.valpha === 1 ? oldHex.call(this) : this.hexa();
-} as Color['hex'];
+} as ColorValue['hex'];
 
 /**
  * Type-safe name to color mapper. Alias for `color(colorName)`